Ajout création Topo controleur + modif suppression Topo et Spot
Quand on supprime un spot ou un topo, on supprimer aussi les photos du serveur.
Quand on supprime un Topo, on supprime aussi les topoUser associés.

+++ src/main/java/com/openclassrooms/escalade/services/SpotService.java
@@ -100,7 +100,9 @@ public class SpotService {
             }
         }
         Spot spot = spotRepository.findById(id).orElseThrow(EntityNotFoundException::new);
-        spot.getPhotos().forEach(photo -> this.fileStorageService.delete(photo.getName()));
+        if (spot.getPhotos() != null && spot.getPhotos().size() > 0) {
+            spot.getPhotos().forEach(photo -> this.fileStorageService.delete(photo.getName()));
+        }
         spotRepository.delete(spot);
     }
